requirements.txt
alembic==1.8.1
asttokens==2.0.5
bcrypt==3.2.2
cffi==1.15.1
click==8.0.4
colorama==0.4.4
executing==0.8.2
Flask==2.2.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
flask-optimize==0.2.9.8
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
htmlmin==0.1.12
icecream==2.1.2
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.2.1
Markdown==3.3.6
MarkupSafe==2.1.1
-e git+https://github.com/corei8/MicroSite.git@275c53510a6fc07a9a0c8894ea38d3350f546905#egg=microsite
pathlib==1.0.1
pycparser==2.21
Pygments==2.11.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
zipp==3.7.0
